requirements.txt
argcomplete==1.12.3
boto3==1.18.44
botocore==1.21.44
certifi==2021.5.30
cfn-flip==1.2.3
charset-normalizer==2.0.5
click==7.1.2
durationpy==0.5
Flask==1.1.4
Flask-Cors==3.0.10
future==0.18.2
hjson==3.0.2
idna==3.2
importlib-metadata==4.8.1
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
kappa==0.6.0
MarkupSafe==2.0.1
pep517==0.11.0
pip-tools==6.2.0
placebo==0.9.0
PyJWT==2.1.0
python-dateutil==2.6.1
python-slugify==5.0.2
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
tqdm==4.62.2
troposphere<3
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.53.0
zipp==3.5.0
